Branch checkout after cloning (#460)
* Git Clone: Add basic functionality of cloning repository

Basics of cloning repository implemented. Pressing clone on welcome
view now opens simple window with url input with clone and close buttons

For now it clones the repositories in to your home dir. Error handling
for now is nada and should be implemented in later commits

* Git Clone: Select target folder and clone the repository

Now you can select the folder to which you want to clone the repository.
This defaults to home folder.

Also added check that the repository isn't empty when trying to clone.

* Git Clone: Open cloned folder

Now after we press `clone` button, we firstly clone the specified
repository to given folder. After cloning, we open folder picker for
the user to pick folder to open the editor in. However this could be better if
we straight up open the editor in the cloned folder, rather than giving user an option
to choose the folder in between.

* Git Clone: Add error handling

Added error handling for git related errors e.g. when trying to clone
non existent repository

* Git Clone: Implementation as a module

Changed GitClone to be module as discussed in
https://github.com/CodeEditApp/CodeEdit/pull/232#discussion_r834529266

* Git Clone: Uncomment dependency from Package.swift

* Git Clone: Switch to use NSSavePanel

Switched the git clone view folder picking to use NSSavePanel
instead of NSOpenPanel.

NSSavePanel gives us kind of the same flow as Xcode has, as it's also using
NSSavePanel to prompt the target folder for cloning.

The view now consist of two separate modals that pop open, and it may
seem cumbersome, but it now has kind of the same functionality as Xcode.

* Git Clone: Remove .git only if the url has it.

* Git Clone: Reposition UI

Repositioning buttons, input and text. Also added app icon to the view

* Git Clone: Change clone view to be sheet

Changed git clone view to be sheet instead of normal window. This
is better for UX as you can't click anywhere else than the clone view
until you either cancel or clone the repository

* Git Clone: Validate url and more error handling

Naive Xcode-like check for the url implemented. For now if the url
does not start with certain string, the clone button is disabled, just
like in Xcode.

Also added check for when user decides to press cancel on folder selection
https://github.com/CodeEditApp/CodeEdit/pull/232#issuecomment-1081211042

* Git Clone: Add automatic pasting

If you have 'valid' git url in your clipboard when opening the clone
view, it now automatically pastes it to the textfield

* GitClone: Fix some force unwrappings and make vars private

* GitClone: Make shellClient constant

* GitClone: Fix indentation and make logic private func

Fixed indentation in WelcomeView and moved the logic of cloning
repository into it's own private function

* GitClone: Add CheckoutBranchModal

Added modal to checkout to other branches after cloning the repository

* GitClone: Make checkout button defaultAction

* GitClone: Add possibility to checkout to remote branches

Showing all branhes to make it possible to checkout to remote
branches as well as local

* GitClone: Make the modal wider

* GitClone: Open checkout branch modal only if there's multiple branches

* GitClone: Refactor extension to separate file.