Enhancement: #218 Fuzzy search for QuickOpen (#1307)
* Added a fuzzy search for QuickOpen searching;
Every search result gets a score based on how many queryTokes are found in the file url, the proximity Score, the modifying date and the Levenshtein-Distance.

* Added better documentation for some of the FuzzySearch functions;
Fixed the levenshteinDistance function;

* Resolved merge conflicts

* Removed line width warnings;

* Removed redundant else statement

* Removed redundant else statement

* Replaced deprecated methods

* Delete Icon

* Removed unnecessary documentation;
Resolved the bug, where random files weren't displayed;
Spelling mistake corrected;

* Search Results ignore .git files;

* Search Results show .git files again;

* Shortened return statement;

* Found and fixed the missing files error

The problem was the the `calculateDateScore` function, returned a score of 0.0 for files that were never created or modified by the user(downloaded files for example). The date score got multiplied with the file score, and something times 0.0 equals 0 and consequently the items weren't displayed in the list🤦🏻‍♂️. I simply added `score + 0.01` which assigns a default score of 0.01 to each date score. However, there is a disadvantage: the search results are worse...

* Revert "Merge branch 'main' of https://github.com/activcoding/CodeEdit into FuzzySearch"

This reverts commit 42d5440c63962bc70a277ce8a3406700032ee171, reversing
changes made to e67ff21972e61813ed01a5cc6242ac0445348ae2.

* Revert "Revert "Merge branch 'main' of https://github.com/activcoding/CodeEdit into FuzzySearch""

This reverts commit a2108dbcf75dfe527f66ac4c151e2416ff9a7d57.

Signed-off-by: Wouter01 <wouterhennen@gmail.com>

* Fix Xcode issues

Signed-off-by: Wouter01 <wouterhennen@gmail.com>

---------

Signed-off-by: Wouter01 <wouterhennen@gmail.com>
Co-authored-by: Tom Ludwig <83090745+activcoding@users.noreply.github.com>