Fixed terminal resize and scroll layout issues, improved utility area toggle animation (#1845)
When the terminal was resized, it displayed partial lines at the bottom due to a remainder in the line height multiple. This caused two things:

1. Resizing the Utility Area drawer caused a visual stutter in the terminal.
2. Partial lines at the bottom created an unsightly visual artifact.

The terminal height must now be a multiple of a single line’s height to resolve these issues. 

Utility area drawer now has a push/pull animation rather than a reveal.

### Screenshots

Before

https://github.com/user-attachments/assets/f092611e-fe28-4aeb-b4dd-408611f0229e

https://github.com/user-attachments/assets/4c365d3d-e051-4bc4-86ef-7121e4b8a5a5

After

https://github.com/user-attachments/assets/7fb1a8d0-1aac-42f2-8a0a-22c02a7511b0

VS Code handles this in a similar way

https://github.com/user-attachments/assets/67507c80-e39e-45f4-8432-5e132df5116a

Improved utility area animation

https://github.com/user-attachments/assets/36825fd0-caa0-463e-a367-5c17cde7dce2

> [!NOTE]
> I removed the maximize drawer toggle in favor of simply dragging to resize up. This simplifies things greatly.

Co-authored-by: Tom Ludwig <tommludwig@icloud.com>