Open New Files in Workspace (#2043)
### Description

When a new file is opened, tries to open it in a nearby workspace. Right now new files are opened in their own window no matter what (eg using `⌘N` or a file URL).

The workspace is chosen by proximity, so if a user has `/User/Desktop` and `/User/Desktop/Project` open. Opening the file `/User/Desktop/Project/MAKEFILE` will open it in the `Project` workspace.

### Related Issues

* Reported by @ BSM on Discord

### Checklist

- [x] I read and understood the [contributing guide](https://github.com/CodeEditApp/CodeEdit/blob/main/CONTRIBUTING.md) as well as the [code of conduct](https://github.com/CodeEditApp/CodeEdit/blob/main/CODE_OF_CONDUCT.md)
- [x] The issues this PR addresses are related to each other
- [x] My changes generate no new warnings
- [x] My code builds and runs on my machine
- [x] My changes are all related to the related issue above
- [x] I documented my code

### Screenshots

(please ignore the messed up gutter, it's been fixed in CESE but not cut in a release)

https://github.com/user-attachments/assets/8d341dee-d043-4b0a-8fd3-f544bf952abb

Files open to the nearest folder, and bring the window to front.

https://github.com/user-attachments/assets/96ec98a5-20ec-4d53-a996-446d9264e7c4