Implement Invisible Characters Setting (#2065)
### Description

Implements invisible character drawing in CodeEdit, including warning characters for ambiguous or invisible characters.

This is stored in a new submenu in the Text Editing settings page. I decided a pane like this was the best option, as a table cannot grow to fit it's contents or scroll when it's placed in some Forms (like our settings pages for some reason). The table also takes up a lot of space, so I felt it was a good use of a sheet.

### Related Issues

* closes https://github.com/CodeEditApp/CodeEditTextView/issues/22
* closes https://github.com/CodeEditApp/CodeEditSourceEditor/issues/207

### Checklist

- [x] I read and understood the [contributing guide](https://github.com/CodeEditApp/CodeEdit/blob/main/CONTRIBUTING.md) as well as the [code of conduct](https://github.com/CodeEditApp/CodeEdit/blob/main/CODE_OF_CONDUCT.md)
- [x] The issues this PR addresses are related to each other
- [x] My changes generate no new warnings
- [x] My code builds and runs on my machine
- [x] My changes are all related to the related issue above
- [x] I documented my code

### Screenshots

https://github.com/user-attachments/assets/9d3b59c8-c506-471d-8ab6-840860e0c526