Adjust Git Status Parsing to Better Handle Null Chars (#2132)
### Description

Fixes an issue with git status parsing where the substring indexing was
off-by-one when creating the file string due to an index being
incremented before creating a substring.

Fixes this by incrementing the current string index *after* returning
the correct substring.

The related issue is marked as a Tahoe bug, the Tahoe bug is that Tahoe
now shows the null character as a %00 at the end of the label, where
previous macOS versions did not. This change is retroactive however as
it is a bug fix.

### Related Issues

* closes #2119 

### Checklist

- [x] I read and understood the [contributing
guide](https://github.com/CodeEditApp/CodeEdit/blob/main/CONTRIBUTING.md)
as well as the [code of
conduct](https://github.com/CodeEditApp/CodeEdit/blob/main/CODE_OF_CONDUCT.md)
- [x] The issues this PR addresses are related to each other
- [x] My changes generate no new warnings
- [x] My code builds and runs on my machine
- [x] My changes are all related to the related issue above
- [x] I documented my code

### Screenshots

N/A